requirements.txt
# Core Frameworks
langchain==0.3.25           
fastapi==0.115.12         
uvicorn[standard]==0.34.2   
# Embeddings & Vector Store
sentence-transformers==4.1.0 
faiss-cpu==1.11.0             
# LLM & Hugging Face Integration
huggingface_hub==0.30.2      
langchain-huggingface==0.1.2 
# Required by Flan-T5
transformers==4.51.3         
torch==2.4.1                 
accelerate==1.6.0           

# Environment Variables
python-dotenv==1.1.0         

# Optional: For document loading (if transcripts are PDF, etc.)
pypdf==5.4.0               

langchain-community

protobuf==3.20.3
# ---
# Note: Versions are based on search results and PyPI checks around early May 2024/2025.
# Always verify latest versions on PyPI (pypi.org) before finalizing dependencies.
# Pinning versions ensures environment reproducibility.
